# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PlanejaPro API - Copiloto inteligente para planejamento pedagógico
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface PlanningInput {
  /** Disciplina (ex. Matemática, Português) */
  disciplina: string;
  /** Ano/Série (ex. 6º Ano, 2º EM) */
  anoSerie: string;
  /** Identificação da turma */
  turma?: string;
  /**
     * Número de aulas no planejamento
     * @minimum 1
     * @maximum 50
     */
  quantidadeAulas: number;
  /**
     * Duração de cada aula em minutos
     * @minimum 30
     * @maximum 240
     */
  duracaoAula: number;
  /** Perfis da turma (participativa, agitada, etc) */
  perfilTurma?: string[];
  /** Objetivos desejados pelo professor */
  objetivos?: string;
  /** Conteúdo a ser trabalhado */
  conteudo: string;
  /** Recursos disponíveis (projetor, internet, etc) */
  recursosDisponiveis?: string[];
  /** Observações adicionais do professor */
  observacoes?: string;
}

export interface GeneratedPlanning {
  tema: string;
  objetivoGeral: string;
  objetivosEspecificos: string[];
  competencias: string[];
  habilidades: string[];
  metodologia: string;
  sequenciaDidatica: string[];
  atividadeInicial: string;
  desenvolvimento: string;
  atividadePratica: string;
  encerramento: string;
  avaliacao: string;
  criteriosAvaliativos: string[];
  estrategiasInclusivas: string;
  adaptacoesDificuldades: string;
  recursosNecessarios: string[];
  tarefaCasa: string;
  observacoesPedagogicas: string;
  versaoResumida: string;
  sugestoesExtras: string[];
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface ChatInput {
  /** Mensagem do professor ao assistente */
  message: string;
  /** Histórico de mensagens anteriores */
  history?: ChatMessage[];
  /** Contexto do planejamento atual (opcional) */
  planningContext?: string;
}

export interface ChatResponse {
  /** Resposta do assistente pedagógico */
  message: string;
}