# lib/api-zod/src/generated/types/chatInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PlanejaPro API - Copiloto inteligente para planejamento pedagógico
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from './chatMessage';

export interface ChatInput {
  /** Mensagem do professor ao assistente */
  message: string;
  /** Histórico de mensagens anteriores */
  history?: ChatMessage[];
  /** Contexto do planejamento atual (opcional) */
  planningContext?: string;
}