# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PlanejaPro API - Copiloto inteligente para planejamento pedagógico
 * OpenAPI spec version: 0.1.0
 */

export interface ChatResponse {
  /** Resposta do assistente pedagógico */
  message: string;
}