# lib/api-zod/src/generated/types/generatedPlanning.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PlanejaPro API - Copiloto inteligente para planejamento pedagógico
 * OpenAPI spec version: 0.1.0
 */

export interface GeneratedPlanning {
  tema: string;
  objetivoGeral: string;
  objetivosEspecificos: string[];
  competencias: string[];
  habilidades: string[];
  metodologia: string;
  sequenciaDidatica: string[];
  atividadeInicial: string;
  desenvolvimento: string;
  atividadePratica: string;
  encerramento: string;
  avaliacao: string;
  criteriosAvaliativos: string[];
  estrategiasInclusivas: string;
  adaptacoesDificuldades: string;
  recursosNecessarios: string[];
  tarefaCasa: string;
  observacoesPedagogicas: string;
  versaoResumida: string;
  sugestoesExtras: string[];
}