# lib/api-zod/src/generated/types/planningInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PlanejaPro API - Copiloto inteligente para planejamento pedagógico
 * OpenAPI spec version: 0.1.0
 */

export interface PlanningInput {
  /** Disciplina (ex. Matemática, Português) */
  disciplina: string;
  /** Ano/Série (ex. 6º Ano, 2º EM) */
  anoSerie: string;
  /** Identificação da turma */
  turma?: string;
  /**
     * Número de aulas no planejamento
     * @minimum 1
     * @maximum 50
     */
  quantidadeAulas: number;
  /**
     * Duração de cada aula em minutos
     * @minimum 30
     * @maximum 240
     */
  duracaoAula: number;
  /** Perfis da turma (participativa, agitada, etc) */
  perfilTurma?: string[];
  /** Objetivos desejados pelo professor */
  objetivos?: string;
  /** Conteúdo a ser trabalhado */
  conteudo: string;
  /** Recursos disponíveis (projetor, internet, etc) */
  recursosDisponiveis?: string[];
  /** Observações adicionais do professor */
  observacoes?: string;
}